ASM-4248 Zone configuration is skipping server's WWPN. And zone is not added to the zoneset
Due to recent refactoring changes, transport is getting initiatlized for each of the resource invocation. There is a limit on the number of connections that we can open on Brocade switch (2).
To take care of the connection pool, added connection close call (transport.close) in create and destroy methods of the resources used by ASM